Add Transaction for Reservation Procces

+++ Domain/Interfaces/ISpaceRepository.cs
@@ -5,7 +5,7 @@ namespace Domain.Interfaces;
 public interface ISpaceRepository
 {
     Task AddSpace(AddSpaceCommand addSpaceCommand);
-    Task<double> GetPriceOfSpace(int SpaceId);
+    Task<int> GetPriceOfSpace(int SpaceId);
     Task<int> SpaceIsFreeAtDate(int SpaceId, DateOnly dateOnly);
 
 }
